[opt] use secondary index when join on indexed columns
