+++ pkg/sql/plan/deepcopy.go
@@ -16,6 +16,7 @@ package plan
 
 import (
 	"bytes"
+
 	"github.com/matrixorigin/matrixone/pkg/pb/plan"
 	"golang.org/x/exp/constraints"
 )
@@ -979,9 +980,7 @@ func DeepCopyExpr(expr *Expr) *Expr {
 
 	case *plan.Expr_T:
 		newExpr.Expr = &plan.Expr_T{
-			T: &plan.TargetType{
-				Typ: DeepCopyType(item.T.Typ),
-			},
+			T: &plan.TargetType{},
 		}
 
 	case *plan.Expr_Max:
