+++ pkg/sql/plan/function/init.go
@@ -24,6 +24,8 @@ var AggSumOverloadID = encodeOverloadID(SUM, 0)
 
 var InFunctionEncodedID = encodeOverloadID(IN, 0)
 var InFunctionName = "in"
+var PrefixInFunctionEncodedID = encodeOverloadID(PREFIX_IN, 0)
+var PrefixInFunctionName = "prefix_in"
 var L2DistanceFunctionEncodedID = encodeOverloadID(L2_DISTANCE, 0)
 
 func init() {
