+++ pkg/sql/plan/opt_misc.go
@@ -201,483 +201,6 @@ func replaceColumnsForExpr(expr *plan.Expr, projMap map[[2]int32]*plan.Expr) *pl
 	return expr
 }
 
-func (builder *QueryBuilder) pushdownFilters(nodeID int32, filters []*plan.Expr, separateNonEquiConds bool) (int32, []*plan.Expr) {
-	node := builder.qry.Nodes[nodeID]
-
-	var canPushdown, cantPushdown []*plan.Expr
-
-	if node.Limit != nil {
-		// can not push down over limit
-		cantPushdown = filters
-		filters = nil
-	}
-
-	switch node.NodeType {
-	case plan.Node_AGG:
-		groupTag := node.BindingTags[0]
-		aggregateTag := node.BindingTags[1]
-
-		for _, filter := range filters {
-			if !containsTag(filter, aggregateTag) {
-				canPushdown = append(canPushdown, replaceColRefs(filter, groupTag, node.GroupBy))
-			} else {
-				node.FilterList = append(node.FilterList, filter)
-			}
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-	case plan.Node_SAMPLE:
-		groupTag := node.BindingTags[0]
-		sampleTag := node.BindingTags[1]
-
-		for _, filter := range filters {
-			if !containsTag(filter, sampleTag) {
-				canPushdown = append(canPushdown, replaceColRefs(filter, groupTag, node.GroupBy))
-			} else {
-				node.FilterList = append(node.FilterList, filter)
-			}
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-	case plan.Node_WINDOW:
-		windowTag := node.BindingTags[0]
-
-		for _, filter := range filters {
-			if !containsTag(filter, windowTag) {
-				canPushdown = append(canPushdown, replaceColRefs(filter, windowTag, node.WinSpecList))
-			} else {
-				node.FilterList = append(node.FilterList, filter)
-			}
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-	case plan.Node_TIME_WINDOW:
-		windowTag := node.BindingTags[0]
-
-		for _, filter := range filters {
-			if !containsTag(filter, windowTag) {
-				canPushdown = append(canPushdown, replaceColRefs(filter, windowTag, node.WinSpecList))
-			} else {
-				node.FilterList = append(node.FilterList, filter)
-			}
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-	case plan.Node_FILTER:
-		canPushdown = filters
-		for _, filter := range node.FilterList {
-			canPushdown = append(canPushdown, splitPlanConjunction(applyDistributivity(builder.GetContext(), filter))...)
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			node.Children[0] = childID
-			node.FilterList = cantPushdownChild
-		} else {
-			nodeID = childID
-		}
-
-	case plan.Node_JOIN:
-		leftTags := make(map[int32]emptyType)
-		for _, tag := range builder.enumerateTags(node.Children[0]) {
-			leftTags[tag] = emptyStruct
-		}
-
-		rightTags := make(map[int32]emptyType)
-		for _, tag := range builder.enumerateTags(node.Children[1]) {
-			rightTags[tag] = emptyStruct
-		}
-
-		var markTag int32
-		if node.JoinType == plan.Node_MARK {
-			markTag = node.BindingTags[0]
-		}
-
-		node.OnList = splitPlanConjunctions(node.OnList)
-
-		if node.JoinType == plan.Node_INNER {
-			for _, cond := range node.OnList {
-				filters = append(filters, splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))...)
-			}
-
-			node.OnList = nil
-		}
-
-		var leftPushdown, rightPushdown []*plan.Expr
-		var turnInner bool
-
-		joinSides := make([]int8, len(filters))
-
-		for i, filter := range filters {
-			canTurnInner := true
-
-			joinSides[i] = getJoinSide(filter, leftTags, rightTags, markTag)
-			if f, ok := filter.Expr.(*plan.Expr_F); ok {
-				for _, arg := range f.F.Args {
-					if getJoinSide(arg, leftTags, rightTags, markTag) == JoinSideBoth {
-						canTurnInner = false
-						break
-					}
-				}
-			}
-
-			if canTurnInner && node.JoinType == plan.Node_LEFT && joinSides[i]&JoinSideRight != 0 && rejectsNull(filter, builder.compCtx.GetProcess()) {
-				for _, cond := range node.OnList {
-					filters = append(filters, splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))...)
-				}
-
-				node.JoinType = plan.Node_INNER
-				node.OnList = nil
-				turnInner = true
-
-				break
-			}
-
-			// TODO: FULL OUTER join should be handled here. However we don't have FULL OUTER join now.
-		}
-
-		if turnInner {
-			joinSides = make([]int8, len(filters))
-
-			for i, filter := range filters {
-				joinSides[i] = getJoinSide(filter, leftTags, rightTags, markTag)
-			}
-		} else if node.JoinType == plan.Node_LEFT {
-			var newOnList []*plan.Expr
-			for _, cond := range node.OnList {
-				conj := splitPlanConjunction(applyDistributivity(builder.GetContext(), cond))
-				for _, conjElem := range conj {
-					side := getJoinSide(conjElem, leftTags, rightTags, markTag)
-					if side&JoinSideLeft == 0 {
-						rightPushdown = append(rightPushdown, conjElem)
-					} else {
-						newOnList = append(newOnList, conjElem)
-					}
-				}
-			}
-
-			node.OnList = newOnList
-		}
-
-		if !separateNonEquiConds {
-			var extraFilters []*plan.Expr
-			for i, filter := range filters {
-				if joinSides[i] != JoinSideBoth {
-					continue
-				}
-				switch exprImpl := filter.Expr.(type) {
-				case *plan.Expr_F:
-					if exprImpl.F.Func.ObjName == "or" {
-						keys := checkDNF(filter)
-						for _, key := range keys {
-							extraFilter := walkThroughDNF(builder.GetContext(), filter, key)
-							if extraFilter != nil {
-								extraFilters = append(extraFilters, DeepCopyExpr(extraFilter))
-								joinSides = append(joinSides, getJoinSide(extraFilter, leftTags, rightTags, markTag))
-							}
-						}
-					}
-				}
-			}
-			filters = append(filters, extraFilters...)
-		}
-
-		for i, filter := range filters {
-			switch joinSides[i] {
-			case JoinSideNone:
-				if c, ok := filter.Expr.(*plan.Expr_Lit); ok {
-					if c, ok := c.Lit.Value.(*plan.Literal_Bval); ok {
-						if c.Bval {
-							break
-						}
-					}
-				}
-
-				switch node.JoinType {
-				case plan.Node_INNER:
-					leftPushdown = append(leftPushdown, DeepCopyExpr(filter))
-					rightPushdown = append(rightPushdown, filter)
-
-				case plan.Node_LEFT, plan.Node_SEMI, plan.Node_ANTI, plan.Node_SINGLE, plan.Node_MARK:
-					leftPushdown = append(leftPushdown, filter)
-
-				default:
-					cantPushdown = append(cantPushdown, filter)
-				}
-
-			case JoinSideLeft:
-				if node.JoinType != plan.Node_OUTER {
-					leftPushdown = append(leftPushdown, filter)
-				} else {
-					cantPushdown = append(cantPushdown, filter)
-				}
-
-			case JoinSideRight:
-				if node.JoinType == plan.Node_INNER {
-					rightPushdown = append(rightPushdown, filter)
-				} else {
-					cantPushdown = append(cantPushdown, filter)
-				}
-
-			case JoinSideBoth:
-				if node.JoinType == plan.Node_INNER {
-					if separateNonEquiConds {
-						if f, ok := filter.Expr.(*plan.Expr_F); ok {
-							if f.F.Func.ObjName == "=" {
-								if getJoinSide(f.F.Args[0], leftTags, rightTags, markTag) != JoinSideBoth {
-									if getJoinSide(f.F.Args[1], leftTags, rightTags, markTag) != JoinSideBoth {
-										node.OnList = append(node.OnList, filter)
-										break
-									}
-								}
-							}
-						}
-					} else {
-						node.OnList = append(node.OnList, filter)
-						break
-					}
-				}
-
-				cantPushdown = append(cantPushdown, filter)
-
-			case JoinSideMark:
-				if tryMark, ok := filter.Expr.(*plan.Expr_Col); ok {
-					if tryMark.Col.RelPos == node.BindingTags[0] {
-						node.JoinType = plan.Node_SEMI
-						node.BindingTags = nil
-						break
-					}
-				} else if fExpr, ok := filter.Expr.(*plan.Expr_F); ok {
-					if filter.Typ.NotNullable && fExpr.F.Func.ObjName == "not" {
-						arg := fExpr.F.Args[0]
-						if tryMark, ok := arg.Expr.(*plan.Expr_Col); ok {
-							if tryMark.Col.RelPos == node.BindingTags[0] {
-								node.JoinType = plan.Node_ANTI
-								node.BindingTags = nil
-								break
-							}
-						}
-					}
-				}
-
-				cantPushdown = append(cantPushdown, filter)
-
-			default:
-				cantPushdown = append(cantPushdown, filter)
-			}
-		}
-
-		//when onlist is empty, it will be a cross join, performance will be very poor
-		//in this situation, we put the non equal conds in the onlist and go loop join
-		//todo: when equal conds and non equal conds both exists, put them in the on list and go hash equal join
-		if node.JoinType == plan.Node_INNER && len(node.OnList) == 0 {
-			// for tpch q22, do not change the plan for now. will fix in the future
-			leftStats := builder.qry.Nodes[node.Children[0]].Stats
-			rightStats := builder.qry.Nodes[node.Children[1]].Stats
-			if leftStats.Outcnt != 1 && rightStats.Outcnt != 1 {
-				node.OnList = append(node.OnList, cantPushdown...)
-				cantPushdown = nil
-			}
-		}
-
-		switch node.JoinType {
-		case plan.Node_INNER, plan.Node_SEMI:
-			//inner and semi join can deduce new predicate from both side
-			builder.pushdownFilters(node.Children[0], deduceNewFilterList(rightPushdown, node.OnList), separateNonEquiConds)
-			builder.pushdownFilters(node.Children[1], deduceNewFilterList(leftPushdown, node.OnList), separateNonEquiConds)
-		case plan.Node_RIGHT:
-			//right join can deduce new predicate only from right side to left
-			builder.pushdownFilters(node.Children[0], deduceNewFilterList(rightPushdown, node.OnList), separateNonEquiConds)
-		case plan.Node_LEFT:
-			//left join can deduce new predicate only from left side to right
-			builder.pushdownFilters(node.Children[1], deduceNewFilterList(leftPushdown, node.OnList), separateNonEquiConds)
-		}
-
-		if builder.qry.Nodes[node.Children[1]].NodeType == plan.Node_FUNCTION_SCAN {
-
-			for _, filter := range filters {
-				down := false
-				if builder.checkExprCanPushdown(filter, builder.qry.Nodes[node.Children[0]]) {
-					leftPushdown = append(leftPushdown, DeepCopyExpr(filter))
-					down = true
-				}
-				if builder.checkExprCanPushdown(filter, builder.qry.Nodes[node.Children[1]]) {
-					rightPushdown = append(rightPushdown, DeepCopyExpr(filter))
-					down = true
-				}
-				if !down {
-					cantPushdown = append(cantPushdown, DeepCopyExpr(filter))
-				}
-			}
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], leftPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-		childID, cantPushdownChild = builder.pushdownFilters(node.Children[1], rightPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[1]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[1] = childID
-
-	case plan.Node_UNION, plan.Node_UNION_ALL, plan.Node_MINUS, plan.Node_MINUS_ALL, plan.Node_INTERSECT, plan.Node_INTERSECT_ALL:
-		leftChild := builder.qry.Nodes[node.Children[0]]
-		rightChild := builder.qry.Nodes[node.Children[1]]
-		var canPushDownRight []*plan.Expr
-
-		for _, filter := range filters {
-			canPushdown = append(canPushdown, replaceColRefsForSet(DeepCopyExpr(filter), leftChild.ProjectList))
-			canPushDownRight = append(canPushDownRight, replaceColRefsForSet(filter, rightChild.ProjectList))
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-		node.Children[0] = childID
-
-		childID, cantPushdownChild = builder.pushdownFilters(node.Children[1], canPushDownRight, separateNonEquiConds)
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[1]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-		node.Children[1] = childID
-
-	case plan.Node_PROJECT:
-		child := builder.qry.Nodes[node.Children[0]]
-		if (child.NodeType == plan.Node_VALUE_SCAN || child.NodeType == plan.Node_EXTERNAL_SCAN) && child.RowsetData == nil {
-			cantPushdown = filters
-			break
-		}
-
-		projectTag := node.BindingTags[0]
-
-		for _, filter := range filters {
-			canPushdown = append(canPushdown, replaceColRefs(filter, projectTag, node.ProjectList))
-		}
-
-		childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], canPushdown, separateNonEquiConds)
-
-		if len(cantPushdownChild) > 0 {
-			childID = builder.appendNode(&plan.Node{
-				NodeType:   plan.Node_FILTER,
-				Children:   []int32{node.Children[0]},
-				FilterList: cantPushdownChild,
-			}, nil)
-		}
-
-		node.Children[0] = childID
-
-	case plan.Node_TABLE_SCAN, plan.Node_EXTERNAL_SCAN:
-		for _, filter := range filters {
-			if onlyContainsTag(filter, node.BindingTags[0]) {
-				node.FilterList = append(node.FilterList, filter)
-			} else {
-				cantPushdown = append(cantPushdown, filter)
-			}
-		}
-	case plan.Node_FUNCTION_SCAN:
-		downFilters := make([]*plan.Expr, 0)
-		selfFilters := make([]*plan.Expr, 0)
-		for _, filter := range filters {
-			if onlyContainsTag(filter, node.BindingTags[0]) {
-				selfFilters = append(selfFilters, DeepCopyExpr(filter))
-			} else {
-				downFilters = append(downFilters, DeepCopyExpr(filter))
-			}
-		}
-		node.FilterList = append(node.FilterList, selfFilters...)
-		childId := node.Children[0]
-		childId, _ = builder.pushdownFilters(childId, downFilters, separateNonEquiConds)
-		node.Children[0] = childId
-	default:
-		if len(node.Children) > 0 {
-			childID, cantPushdownChild := builder.pushdownFilters(node.Children[0], filters, separateNonEquiConds)
-
-			if len(cantPushdownChild) > 0 {
-				childID = builder.appendNode(&plan.Node{
-					NodeType:   plan.Node_FILTER,
-					Children:   []int32{node.Children[0]},
-					FilterList: cantPushdownChild,
-				}, nil)
-			}
-
-			node.Children[0] = childID
-		} else {
-			cantPushdown = filters
-		}
-	}
-
-	return nodeID, cantPushdown
-}
-
 func (builder *QueryBuilder) swapJoinChildren(nodeID int32) {
 	node := builder.qry.Nodes[nodeID]
 
@@ -729,7 +252,7 @@ func (builder *QueryBuilder) remapWindowClause(expr *plan.Expr, windowTag int32,
 }
 
 /*
-func getJoinCondLeftCol(cond *Expr, leftTags map[int32]emptyType) *plan.Expr_Col {
+func getJoinCondLeftCol(cond *Expr, leftTags map[int32]bool) *plan.Expr_Col {
 	fun, ok := cond.Expr.(*plan.Expr_F)
 	if !ok || fun.F.Func.ObjName != "=" {
 		return nil
@@ -742,10 +265,10 @@ func getJoinCondLeftCol(cond *Expr, leftTags map[int32]emptyType) *plan.Expr_Col
 	if !ok {
 		return nil
 	}
-	if _, ok := leftTags[leftCol.Col.RelPos]; ok {
+	if leftTags[leftCol.Col.RelPos] {
 		return leftCol
 	}
-	if _, ok := leftTags[rightCol.Col.RelPos]; ok {
+	if leftTags[rightCol.Col.RelPos] {
 		return rightCol
 	}
 	return nil
@@ -911,14 +434,14 @@ func determineHashOnPK(nodeID int32, builder *QueryBuilder) {
 		return
 	}
 
-	leftTags := make(map[int32]emptyType)
+	leftTags := make(map[int32]bool)
 	for _, tag := range builder.enumerateTags(node.Children[0]) {
-		leftTags[tag] = emptyStruct
+		leftTags[tag] = true
 	}
 
-	rightTags := make(map[int32]emptyType)
+	rightTags := make(map[int32]bool)
 	for _, tag := range builder.enumerateTags(node.Children[1]) {
-		rightTags[tag] = emptyStruct
+		rightTags[tag] = true
 	}
 
 	exprs := make([]*plan.Expr, 0)
@@ -965,14 +488,14 @@ func getHashColsNDVRatio(nodeID int32, builder *QueryBuilder) float64 {
 	}
 	result := getHashColsNDVRatio(builder.qry.Nodes[node.Children[1]].NodeId, builder)
 
-	leftTags := make(map[int32]emptyType)
+	leftTags := make(map[int32]bool)
 	for _, tag := range builder.enumerateTags(node.Children[0]) {
-		leftTags[tag] = emptyStruct
+		leftTags[tag] = true
 	}
 
-	rightTags := make(map[int32]emptyType)
+	rightTags := make(map[int32]bool)
 	for _, tag := range builder.enumerateTags(node.Children[1]) {
-		rightTags[tag] = emptyStruct
+		rightTags[tag] = true
 	}
 
 	exprs := make([]*plan.Expr, 0)
@@ -1027,56 +550,6 @@ func checkExprInTags(expr *plan.Expr, tags []int32) bool {
 		}
 	}
 	return false
-}
-
-// order by limit can be pushed down to left join
-func (builder *QueryBuilder) pushTopDownToLeftJoin(nodeID int32) {
-	node := builder.qry.Nodes[nodeID]
-	var joinnode, nodePushDown *plan.Node
-	var tags []int32
-	var newNodeID int32
-
-	if node.NodeType != plan.Node_SORT || node.Limit == nil {
-		goto END
-	}
-	joinnode = builder.qry.Nodes[node.Children[0]]
-	if joinnode.NodeType != plan.Node_JOIN {
-		goto END
-	}
-
-	//before join order, only left join
-	if joinnode.JoinType != plan.Node_LEFT {
-		goto END
-	}
-
-	// check orderby column
-	tags = builder.enumerateTags(builder.qry.Nodes[joinnode.Children[0]].NodeId)
-	for i := range node.OrderBy {
-		if !checkExprInTags(node.OrderBy[i].Expr, tags) {
-			goto END
-		}
-	}
-
-	nodePushDown = DeepCopyNode(node)
-
-	if nodePushDown.Offset != nil {
-		newExpr, err := bindFuncExprAndConstFold(builder.GetContext(), builder.compCtx.GetProcess(), "+", []*Expr{nodePushDown.Limit, nodePushDown.Offset})
-		if err != nil {
-			goto END
-		}
-		nodePushDown.Offset = nil
-		nodePushDown.Limit = newExpr
-	}
-	newNodeID = builder.appendNode(nodePushDown, nil)
-	nodePushDown.Children[0] = joinnode.Children[0]
-	joinnode.Children[0] = newNodeID
-
-END:
-	if len(node.Children) > 0 {
-		for _, child := range node.Children {
-			builder.pushTopDownToLeftJoin(child)
-		}
-	}
 }
 
 func (builder *QueryBuilder) rewriteDistinctToAGG(nodeID int32) {
@@ -1139,18 +612,4 @@ func (builder *QueryBuilder) rewriteEffectlessAggToProject(nodeID int32) {
 	node.BindingTags = node.BindingTags[:1]
 	node.ProjectList = node.GroupBy
 	node.GroupBy = nil
-}
-
-func (builder *QueryBuilder) pushdownLimit(nodeID int32) {
-	node := builder.qry.Nodes[nodeID]
-	for _, childID := range node.Children {
-		builder.pushdownLimit(childID)
-	}
-	if node.NodeType == plan.Node_PROJECT && len(node.Children) > 0 {
-		child := builder.qry.Nodes[node.Children[0]]
-		if child.NodeType == plan.Node_TABLE_SCAN {
-			child.Limit, child.Offset = node.Limit, node.Offset
-			node.Limit, node.Offset = nil, nil
-		}
-	}
 }
