+++ pkg/sql/plan/types.go
@@ -155,10 +155,6 @@ type ExecInfo struct {
 	CnNumbers  int
 }
 
-type emptyType struct{}
-
-var emptyStruct = emptyType{}
-
 type QueryBuilder struct {
 	qry     *plan.Query
 	compCtx CompilerContext
@@ -183,7 +179,7 @@ type CTERef struct {
 	defaultDatabase string
 	isRecursive     bool
 	ast             *tree.CTE
-	maskedCTEs      map[string]emptyType
+	maskedCTEs      map[string]bool
 }
 
 type aliasItem struct {
@@ -195,7 +191,7 @@ type BindContext struct {
 	binder Binder
 
 	cteByName              map[string]*CTERef
-	maskedCTEs             map[string]emptyType
+	maskedCTEs             map[string]bool
 	normalCTE              bool
 	initSelect             bool
 	recSelect              bool
