+++ pkg/sql/plan/utils.go
@@ -49,17 +49,17 @@ func GetBindings(expr *plan.Expr) []int32 {
 	return bindings
 }
 
-func doGetBindings(expr *plan.Expr) map[int32]emptyType {
-	res := make(map[int32]emptyType)
+func doGetBindings(expr *plan.Expr) map[int32]bool {
+	res := make(map[int32]bool)
 
 	switch expr := expr.Expr.(type) {
 	case *plan.Expr_Col:
-		res[expr.Col.RelPos] = emptyStruct
+		res[expr.Col.RelPos] = true
 
 	case *plan.Expr_F:
 		for _, child := range expr.F.Args {
 			for id := range doGetBindings(child) {
-				res[id] = emptyStruct
+				res[id] = true
 			}
 		}
 	}
@@ -189,7 +189,7 @@ func decreaseDepth(expr *plan.Expr) (*plan.Expr, bool) {
 	return expr, correlated
 }
 
-func getJoinSide(expr *plan.Expr, leftTags, rightTags map[int32]emptyType, markTag int32) (side int8) {
+func getJoinSide(expr *plan.Expr, leftTags, rightTags map[int32]bool, markTag int32) (side int8) {
 	switch exprImpl := expr.Expr.(type) {
 	case *plan.Expr_F:
 		for _, arg := range exprImpl.F.Args {
@@ -197,9 +197,9 @@ func getJoinSide(expr *plan.Expr, leftTags, rightTags map[int32]emptyType, markT
 		}
 
 	case *plan.Expr_Col:
-		if _, ok := leftTags[exprImpl.Col.RelPos]; ok {
+		if leftTags[exprImpl.Col.RelPos] {
 			side = JoinSideLeft
-		} else if _, ok := rightTags[exprImpl.Col.RelPos]; ok {
+		} else if rightTags[exprImpl.Col.RelPos] {
 			side = JoinSideRight
 		} else if exprImpl.Col.RelPos == markTag {
 			side = JoinSideMark
@@ -800,10 +800,10 @@ func increaseRefCnt(expr *plan.Expr, inc int, colRefCnt map[[2]int32]int) {
 	}
 }
 
-func getHyperEdgeFromExpr(expr *plan.Expr, leafByTag map[int32]int32, hyperEdge map[int32]emptyType) {
+func getHyperEdgeFromExpr(expr *plan.Expr, leafByTag map[int32]int32, hyperEdge map[int32]bool) {
 	switch exprImpl := expr.Expr.(type) {
 	case *plan.Expr_Col:
-		hyperEdge[leafByTag[exprImpl.Col.RelPos]] = emptyStruct
+		hyperEdge[leafByTag[exprImpl.Col.RelPos]] = true
 
 	case *plan.Expr_F:
 		for _, arg := range exprImpl.F.Args {
@@ -1892,7 +1892,7 @@ func HasFkSelfReferOnly(tableDef *TableDef) bool {
 	return true
 }
 
-func MakeInExpr(ctx context.Context, left *Expr, length int32, data []byte) *Expr {
+func MakeInExpr(ctx context.Context, left *Expr, length int32, data []byte, matchPrefix bool) *Expr {
 	rightArg := &plan.Expr{
 		Typ: &plan.Type{
 			Id: int32(types.T_tuple),
@@ -1905,11 +1905,16 @@ func MakeInExpr(ctx context.Context, left *Expr, length int32, data []byte) *Exp
 		},
 	}
 
-	fid := function.InFunctionEncodedID
+	funcID := function.InFunctionEncodedID
+	funcName := function.InFunctionName
+	if matchPrefix {
+		funcID = function.PrefixInFunctionEncodedID
+		funcName = function.PrefixInFunctionName
+	}
 	args := []types.Type{makeTypeByPlan2Expr(left), makeTypeByPlan2Expr(rightArg)}
-	fGet, err := function.GetFunctionByName(ctx, "in", args)
+	fGet, err := function.GetFunctionByName(ctx, funcName, args)
 	if err == nil {
-		fid = fGet.GetEncodedOverloadID()
+		funcID = fGet.GetEncodedOverloadID()
 	}
 	inExpr := &plan.Expr{
 		Typ: &plan.Type{
@@ -1919,8 +1924,8 @@ func MakeInExpr(ctx context.Context, left *Expr, length int32, data []byte) *Exp
 		Expr: &plan.Expr_F{
 			F: &plan.Function{
 				Func: &plan.ObjectRef{
-					Obj:     fid,
-					ObjName: function.InFunctionName,
+					Obj:     funcID,
+					ObjName: funcName,
 				},
 				Args: []*plan.Expr{
 					left,
